uv-resolver: add new ConflictMarker type
This is strictly a refactor. It doesn't change any behavior.

The purpose of this refactor is to make room for supporting more complex
conflict markers. Namely, we need to incorporate package names and then
eventually group names. But we also need sophisticated boolean algebraic
simplification, and the only (cheap, in terms of development cost) to do
that at the moment is to encode our semantics into PEP 508 markers. So
this type will own that encapsulation. (Right now, there's no real
translation happening because we only support extra names, which PEP 508
already supports.)